PKSP-46 run script major refactoring

+++ scripts/pks.sh
@@ -10,8 +10,7 @@ function log () { printf "%b\n" "$*"; }
 function debug () { [[ "$PKS_DEBUG" != "1" ]] || printf "%b\n" "$*" >&2 ; }
 function fail () { log "\nERROR: $*\n" >&2 ; exit 1 ; }
 
-function checkDependencies ()
-{
+function checkDependencies () {
   log "Checking dependencies..."
 
   which curl >/dev/null 2>&1 || fail "Could not find 'curl' command, make sure it's available first before continuing installation."
@@ -26,40 +25,49 @@ function checkDependencies ()
   log ""
 }
 
-#parseParams()
-#{
-#  PKS_VERSION=${1:-"develop"}
-#  PKS_FLAVOR=${2:-"regtest"}
-#
-#
-#  debug "Version: '$PKS_VERSION'"
-#  debug "Flavor: '$PKS_FLAVOR'"
-#  debug "Directory: '$PKS_DIR'"
-#  debug ""
-#}
+function readArgument_PKS_FLAVOR () {
+  PKS_FLAVOR="$2"
 
-function ensureProjectDir () {
+  if [ "$PKS_FLAVOR" == "" ]; then
+    fail "Must specify flavor as a script argument."
+  fi
+}
+
+function ensureWorkDir () {
   if [ ! -d "$PKS_DIR" ]; then
-    log "Project dir '$PKS_DIR' not found, creating project dir..."
+    log "Directory '$PKS_DIR' not found, creating directory..."
 
     mkdir -p "$PKS_DIR"
 
     log "Done."
     log ""
   fi
+
+  log "Running in '$PKS_DIR'."
+  log ""
 }
 
-function downloadDockerComposeFiles () {
-  log "Downloading Docker Compose files into '$PKS_DIR'..."
+function ensureDockerComposeFiles () {
+  log "Searching for '$PKS_VERSION-$PKS_FLAVOR' Docker Compose files in '$PKS_DIR'..."
 
-  __config_file_1_url="https://github.com/PrivateKeySpace/deployment/raw/$PKS_VERSION/config/docker-compose.$PKS_FLAVOR.yml"
   __config_file_1_path="$PKS_DIR/docker-compose.$PKS_FLAVOR.yml"
-
-  curl -fsSL -o "$__config_file_1_path" "$__config_file_1_url" || fail "Failed to download '$__config_file_1_url'."
-
-  __config_file_2_url="https://github.com/PrivateKeySpace/deployment/raw/$PKS_VERSION/config/docker-compose.$PKS_FLAVOR.run.yml"
   __config_file_2_path="$PKS_DIR/docker-compose.$PKS_FLAVOR.run.yml"
 
+  if [ -f "$__config_file_1_path" -a -f "$__config_file_2_path" ]; then
+    log "Found."
+    log ""
+    return
+  fi
+
+  log "Not found."
+  log ""
+
+  log "Downloading '$PKS_VERSION-$PKS_FLAVOR' Docker Compose files into '$PKS_DIR'..."
+
+  __config_file_1_url="https://github.com/PrivateKeySpace/deployment/raw/$PKS_VERSION/config/docker-compose.$PKS_FLAVOR.yml"
+  __config_file_2_url="https://github.com/PrivateKeySpace/deployment/raw/$PKS_VERSION/config/docker-compose.$PKS_FLAVOR.run.yml"
+
+  curl -fsSL -o "$__config_file_1_path" "$__config_file_1_url" || fail "Failed to download '$__config_file_1_url'."
   curl -fsSL -o "$__config_file_2_path" "$__config_file_2_url" || fail "Failed to download '$__config_file_2_url'."
 
   log "Done."
@@ -67,7 +75,7 @@ function downloadDockerComposeFiles () {
 }
 
 function dockerComposePull () {
-  log "\nPulling '$PKS_VERSION-$PKS_FLAVOR' images from Docker Hub...\n"
+  log "Pulling '$PKS_VERSION-$PKS_FLAVOR' images from Docker Hub..."
 
   __config_file_path="$PKS_DIR/docker-compose.$PKS_FLAVOR.yml"
 
@@ -101,24 +109,15 @@ function dockerComposeDown () {
   log ""
 }
 
-function handleInstallCommand () {
-  PKS_FLAVOR="$2"
-
-  downloadDockerComposeFiles
+function handleCommand_install () {
   dockerComposePull
 }
 
-function handleStartCommand () {
-  PKS_FLAVOR="$2"
-
-  downloadDockerComposeFiles
+function handleCommand_start () {
   dockerComposeUp
 }
 
-function handleStopCommand () {
-  PKS_FLAVOR="$2"
-
-  downloadDockerComposeFiles
+function handleCommand_stop () {
   dockerComposeDown
 }
 
@@ -126,18 +125,20 @@ function run () {
   log ""
 
   checkDependencies
+  readArgument_PKS_FLAVOR "$@"
+  ensureDockerComposeFiles
 
   __command="$1"
 
   if [ "$__command" == "install" ]
   then
-      handleInstallCommand
+      handleCommand_install
   elif [ "$__command" == "start" ]
   then
-      handleStartCommand
+      handleCommand_start
   elif [ "$__command" == "stop" ]
   then
-      handleStopCommand
+      handleCommand_stop
   else
       fail "Command '$__command' not supported."
   fi
